requirements.txt
notion-client>=2.0.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
pyinstaller>=6.0.0 